Add `Table.none` (#119)
* Add `Table.none`

Acts as a "black hole" query, rendering as an empty SQL string and not
executed by the database.

* wip

* wip

+++ Sources/StructuredQueriesCore/Statements/Delete.swift
@@ -23,7 +23,7 @@ extension PrimaryKeyedTable {
   /// - Parameter row: A row to delete.
   /// - Returns: A delete statement.
   public static func delete(_ row: Self) -> DeleteOf<Self> {
-    Delete()
+    delete()
       .where {
         $0.primaryKey.eq(TableColumns.PrimaryKey(queryOutput: row[keyPath: $0.primaryKey.keyPath]))
       }
@@ -36,6 +36,7 @@ extension PrimaryKeyedTable {
 ///
 /// To learn more, see <doc:DeleteStatements>.
 public struct Delete<From: Table, Returning> {
+  var isEmpty: Bool
   var `where`: [QueryFragment] = []
   var returning: [QueryFragment] = []
 
@@ -107,6 +108,7 @@ public struct Delete<From: Table, Returning> {
       returning.append("\(quote: resultColumn.name)")
     }
     return Delete<From, (repeat each QueryValue)>(
+      isEmpty: isEmpty,
       where: `where`,
       returning: Array(repeat each selection(From.columns))
     )
@@ -126,6 +128,7 @@ public struct Delete<From: Table, Returning> {
       returning.append("\(quote: resultColumn.name)")
     }
     return Delete<From, From>(
+      isEmpty: isEmpty,
       where: `where`,
       returning: returning
     )
@@ -139,6 +142,7 @@ extension Delete: Statement {
   public typealias QueryValue = Returning
 
   public var query: QueryFragment {
+    guard !isEmpty else { return "" }
     var query: QueryFragment = "DELETE FROM "
     if let schemaName = From.schemaName {
       query.append("\(quote: schemaName).")
